# src/app/api/hr/openapi.yaml
openapi: 3.0.0
info:
  title: HR & Asset Management API
  description: API documentation for the Human Resources & Asset Management module of the Hospital Management System
  version: 1.0.0
  contact:
    name: HMS Support
    email: support@hms.example.com
servers:
  - url: /api
    description: API base URL
...
paths:
  /hr/staff:
    get:
      summary: List employees
      description: Returns a paginated list of employees with optional filtering
      tags:
        - Staff
      parameters:
        - name: skip
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: take
          in: query
          description: Number of records to return per page
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          description: Search term for employee name or ID
          schema:
            type: string
        - name: departmentId
          in: query
          description: Filter by department ID
          schema:
            type: string
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
                  total:
                    type: integer
                  skip:
                    type: integer
                  take:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create employee
      description: Creates a new employee record
      tags:
        - Staff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '200':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /hr/payroll/salary-structures:
    get:
      summary: List salary structures
      description: Returns a paginated list of salary structures with optional filtering
      tags:
        - Payroll
      parameters:
        - name: skip
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: take
          in: query
          description: Number of records to return per page
          schema:
            type: integer
            default: 10
        - name: employeeId
          in: query
          description: Filter by employee ID
          schema:
            type: string
        - name: departmentId
          in: query
          description: Filter by department ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  structures:
                    type: array
                    items:
                      $ref: '#/components/schemas/SalaryStructure'
                  total:
                    type: integer
                  skip:
                    type: integer
                  take:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create salary structure
      description: Creates a new salary structure for an employee
      tags:
        - Payroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalaryStructureInput'
      responses:
        '200':
          description: Salary structure created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryStructure'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        joinDate:
          type: string
          format: date
        terminationDate:
          type: string
          format: date
        isActive:
          type: boolean
        department:
          $ref: '#/components/schemas/Department'
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        qualifications:
          type: array
          items:
            $ref: '#/components/schemas/Qualification'
    EmployeeInput:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        joinDate:
          type: string
          format: date
        departmentId:
          type: string
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [CLINICAL, ADMINISTRATIVE, SUPPORT]
        description:
          type: string
        parentDepartmentId:
          type: string
        parentDepartment:
          $ref: '#/components/schemas/Department'
    Position:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        employeeId:
          type: string
        departmentId:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isPrimary:
          type: boolean
...
    Qualification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [DEGREE, LICENSE, CERTIFICATION, TRAINING]
        name:
          type: string
        issuedBy:
          type: string
        issuedDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        attachmentUrl:
          type: string
        verificationStatus:
          type: string
          enum: [PENDING, VERIFIED, REJECTED]
...
    SalaryStructure:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        employee:
          $ref: '#/components/schemas/Employee'
        baseSalary:
          type: number
          format: float
        effectiveDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        components:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [ALLOWANCE, DEDUCTION, BONUS]
              amount:
                type: number
                format: float
              calculationType:
                type: string
                enum: [FIXED, PERCENTAGE, FORMULA]
              formula:
                type: string
    SalaryStructureInput:
      type: object
      required:
        - employeeId
        - baseSalary
        - effectiveDate
      properties:
        employeeId:
          type: string
        baseSalary:
          type: number
          format: float
        effectiveDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        components:
          type: array
          items:
            type: object
            required:
              - name
              - type
              - amount
              - calculationType
            properties:
              name:
                type: string
              type:
                type: string
                enum: [ALLOWANCE, DEDUCTION, BONUS]
              amount:
                type: number
                format: float
              calculationType:
                type: string
                enum: [FIXED, PERCENTAGE, FORMULA]
              formula:
                type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object